+++ src/main/java/pro/sky/java/course2/calculator/Service.java
@@ -5,5 +5,5 @@ public interface Service {
     String addition(int number1,int number2);
     String subtraction(int number1,int number2);
     String multiplication(int number1,int number2);
-    String division(int number1,int number2);
+    String division(float number1,float number2);
 }
